+++ src/query/storages/fuse/src/operations/read/parquet_data_transform_reader.rs
@@ -40,7 +40,7 @@ use crate::io::TableMetaLocationGenerator;
 use crate::io::VirtualColumnReader;
 use crate::operations::read::block_partition_meta::BlockPartitionMeta;
 use crate::operations::read::data_source_with_meta::DataSourceWithMeta;
-use crate::operations::read::runtime_filter_prunner::runtime_filter_pruner;
+use crate::pruning::ExprRuntimePruner;
 
 pub struct ReadParquetDataTransform<const BLOCKING_IO: bool> {
     func_ctx: FunctionContext,
@@ -124,12 +124,9 @@ impl Transform for ReadParquetDataTransform<true> {
                     self.context
                         .get_min_max_runtime_filter_with_id(self.scan_id),
                 );
-                if runtime_filter_pruner(
-                    self.table_schema.clone(),
-                    &part,
-                    &filters,
-                    &self.func_ctx,
-                )? {
+
+                let runtime_filter = ExprRuntimePruner::new(filters.clone());
+                if runtime_filter.prune(&self.func_ctx, self.table_schema.clone(), &part)? {
                     return Ok(DataBlock::empty());
                 }
 
@@ -207,13 +204,10 @@ impl AsyncTransform for ReadParquetDataTransform<false> {
                             .get_min_max_runtime_filter_with_id(self.scan_id),
                     );
                     let mut fuse_part_infos = Vec::with_capacity(parts.len());
+
+                    let runtime_filter = ExprRuntimePruner::new(filters.clone());
                     for part in parts.into_iter() {
-                        if runtime_filter_pruner(
-                            self.table_schema.clone(),
-                            &part,
-                            &filters,
-                            &self.func_ctx,
-                        )? {
+                        if runtime_filter.prune(&self.func_ctx, self.table_schema.clone(), &part)? {
                             continue;
                         }
